+++ design_bench/oracles/sklearn/gaussian_process_oracle.py
@@ -240,5 +240,5 @@ class GaussianProcessOracle(SKLearnOracle):
         """
 
         # call the model's predict function to generate predictions
-        return self.model.predict(
-            x.reshape((x.shape[0], np.prod(x.shape[1:]))))[:, np.newaxis]
+        return self.model.predict(x.reshape((x.shape[0], np.prod(
+            x.shape[1:]))))[:, np.newaxis].astype(np.float32)
